fixed recommended rout
