+++ src/routes/productItem.route.ts
@@ -5,7 +5,3 @@ import * as productItemController from '../controllers/productItem.controller';
 export const router = express.Router();
 
 router.get('/:id', catchError(productItemController.getOne));
-router.get(
-  '/:id/recommended',
-  catchError(productItemController.getRecommended),
-);
