+++ src/utils/determineRoute.ts
@@ -1,12 +1,10 @@
 import { type Request, type Response, type NextFunction } from 'express';
 
 export const determineRoute = (type: string) => {
-  return (req: Request, _res: Response, next: NextFunction) => {
-    const isID = !isNaN(+req.params?.route);
-
-    if (!isID && type === 'category') {
+  return (_req: Request, _res: Response, next: NextFunction) => {
+    if (type === 'category') {
       next();
-    } else if (isID && type === 'id') {
+    } else if (type === 'itemId') {
       next();
     } else {
       next('route');
